fix(dependencies): downgrade cmdk and next packages to stable versions
This commit updates the dependencies in package.json and pnpm-lock.yaml to provide more stable versions for `cmdk` and `next`. The specific changes include downgrading `cmdk` from version 1.0.4 to 1.0.0 and `next` from version 15.1.7 to 15.1.6.

These downgrades may be necessary to address compatibility issues that arose with the latest versions, ensuring better stability and functionality within the project. Additionally, this update also adjusts the type definitions for Node.js to version 20, ensuring compatibility with recent code practices.